Create package-checker.py
This file checks whether the package "mysql.connector" is present in the system. If not, the code prompts the user for the missing package. if the user declines the code exists and the code will not be able to execute properly, if the user complies, the package is installed by using python's inbuilt package manager 'pip'.